Basic Combat Redo Finished
The basics of the new combat system are finished. Some stats still need
tweaking to get to a good level, but for the most part it is done.

The bug introduced int he previous version has been taken care of.

--BUGS--
The bug where it crashes upon death has reemerged. It has been a fairly
long-lasting one so I'm hoping that I can get it sorted out for good,
soon. I'm really sick of its existence.