Draining blood from corpses
Basic blood well for the player and draining blood from enemy corpses is
in. It still needs a good amount of work to get it fully functioning,
but it IS stable atm.

-- BUGS --
Messages are not properly logged when draining a corpse. It should tell
you when you drain a corpse, and when a corpse is unable to be drained.

CTD on death has not been removed yet.